[GCC] Unreviewed, build fix for Ubuntu 20.04 after 260367@main https://bugs.webkit.org/show_bug.cgi?id=252449
Add static cast from "unsigned long" to "unsigned long long" (supported by
WebGLAny).

* Source/WebCore/html/canvas/EXTDisjointTimerQuery.cpp:
(WebCore::EXTDisjointTimerQuery::getQueryObjectEXT):

Canonical link: https://commits.webkit.org/260428@main

+++ Source/WebCore/html/canvas/EXTDisjointTimerQuery.cpp
@@ -249,7 +249,7 @@ WebGLAny EXTDisjointTimerQuery::getQueryObjectEXT(WebGLTimerQueryEXT& query, GCG
     case GraphicsContextGL::QUERY_RESULT_EXT:
         if (!query.isResultAvailable())
             return 0;
-        return context->graphicsContextGL()->getQueryObjectui64EXT(query.object(), pname);
+        return static_cast<unsigned long long>(context->graphicsContextGL()->getQueryObjectui64EXT(query.object(), pname));
     case GraphicsContextGL::QUERY_RESULT_AVAILABLE_EXT:
         if (!query.isResultAvailable())
             return false;
